Refactoring: Create 'Gender' enum and Move authentication related classes to 'auth' domain
