+++ src/main/java/com/honeypot/domain/auth/entity/enums/AuthProviderType.java
@@ -1,0 +1,5 @@
+package com.honeypot.domain.auth.entity.enums;
+
+public enum AuthProviderType {
+    KAKAO
+}
