+++ src/main/java/com/honeypot/domain/auth/service/KakaoAuthService.java
@@ -5,9 +5,9 @@ import com.honeypot.domain.auth.dto.kakao.KakaoTokenInfo;
 import com.honeypot.domain.auth.dto.kakao.KakaoTokenIssuance;
 import com.honeypot.domain.auth.dto.kakao.KakaoUserInfo;
 import com.honeypot.domain.auth.repository.KakaoAuthRepository;
-import com.honeypot.domain.member.entity.AuthProvider;
+import com.honeypot.domain.auth.entity.AuthProvider;
 import com.honeypot.domain.member.entity.Member;
-import com.honeypot.domain.member.enums.AuthProviderType;
+import com.honeypot.domain.auth.entity.enums.AuthProviderType;
 import com.honeypot.domain.member.repository.AuthProviderRepository;
 import com.honeypot.domain.member.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
