+++ src/main/java/com/honeypot/domain/member/entity/enums/Gender.java
@@ -1,0 +1,5 @@
+package com.honeypot.domain.member.entity.enums;
+
+public enum Gender {
+    MALE, FEMALE
+}
