--- src/main/java/com/honeypot/domain/member/enums/AuthProviderType.java
@@ -1,5 +1,0 @@
-package com.honeypot.domain.member.enums;
-
-public enum AuthProviderType {
-    KAKAO
-}
