+++ src/main/java/com/honeypot/domain/member/repository/AuthProviderRepository.java
@@ -1,7 +1,7 @@
 package com.honeypot.domain.member.repository;
 
-import com.honeypot.domain.member.entity.AuthProvider;
-import com.honeypot.domain.member.enums.AuthProviderType;
+import com.honeypot.domain.auth.entity.AuthProvider;
+import com.honeypot.domain.auth.entity.enums.AuthProviderType;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
